requirements.txt
whatsfly-labfox
openai-whisper